Split out AES encryption package
